+++ SevenZip/EventArguments/ProgressEventArgs.cs
@@ -10,7 +10,7 @@
         /// </summary>
         /// <param name="percentDone">The percent of finished work.</param>
         /// <param name="percentDelta">The percent of work done after the previous event.</param>
-        public ProgressEventArgs(byte percentDone, byte percentDelta)
+        public ProgressEventArgs(float percentDone, float percentDelta)
             : base(percentDone)
         {
             PercentDelta = percentDelta;
@@ -19,6 +19,6 @@
         /// <summary>
         /// Gets the change in done work percentage.
         /// </summary>
-        public byte PercentDelta { get; }
+        public float PercentDelta { get; }
     }
 }
